+++ testSDK/AppListView/DeviceListView.cpp
@@ -121,7 +121,7 @@ void CDeviceListView::showEvent(QShowEvent * e) {
     m_pList->getDeviceList(DeviceList);
 
     if (DeviceList.size() > 0) {
-        for (int i = 0; i < DeviceList.size(); ++i) {
+        for (size_t i = 0; i < DeviceList.size(); ++i) {
             InsertDevice(1+i,DeviceList[i].id,
                            DeviceList[i].name.c_str(),
                            appsheet_on[0],
